# app/server/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Open Enterprise REST API",
    "version": "1.3.1",
    "description": "Developer API for querying, managing, and automating Open Enterprise workspaces programmatically.\n\n## Authentication\nAll endpoints require an API key. Generate one in **Admin Panel → API Keys**.\n\nInclude it in every request as a Bearer token:\n```\nAuthorization: Bearer emb_your_api_key_here\n```\n\n## Base URL\nAll endpoints are relative to your Open Enterprise instance URL, e.g. `https://your-domain.com/api/v1`"
  },
  "servers": [
    { "url": "/api/v1", "description": "This Open Enterprise instance" }
  ],
  "security": [
    { "ApiKeyAuth": [] }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Open Enterprise API Key",
        "description": "Create an API key in Admin Panel → API Keys. Keys start with `emb_`."
      }
    },
    "schemas": {
...
      "Agent": {
        "type": "object",
        "properties": {
          "id":             { "type": "integer", "example": 1 },
          "name":           { "type": "string",  "example": "Daily Report Agent" },
          "triggerType":    { "type": "string",  "enum": ["manual","scheduled"], "example": "scheduled" },
          "cronExpression": { "type": "string",  "example": "0 9 * * *", "nullable": true },
          "enabled":        { "type": "boolean", "example": true },
          "createdAt":      { "type": "string",  "format": "date-time" }
        }
      },
      "AgentRun": {
        "type": "object",
        "properties": {
          "id":          { "type": "integer", "example": 42 },
          "status":      { "type": "string",  "enum": ["running","success","error"], "example": "success" },
          "triggerType": { "type": "string",  "enum": ["manual","scheduled"], "example": "manual" },
          "input":       { "type": "string",  "nullable": true },
          "output":      { "type": "string",  "nullable": true },
          "error":       { "type": "string",  "nullable": true },
          "startedAt":   { "type": "string",  "format": "date-time" },
          "completedAt": { "type": "string",  "format": "date-time", "nullable": true }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Workspace not found" }
        }
      }
    }
  },
  "paths": {
...
    "/workspaces/{slug}/agents": {
      "get": {
        "summary": "List agents",
        "operationId": "listAgents",
        "tags": ["Agents"],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "List of agents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "agents": { "type": "array", "items": { "$ref": "#/components/schemas/Agent" } } }
                }
              }
            }
          },
          "401": { "description": "Invalid or missing API key", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Workspace not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/workspaces/{slug}/agents/{id}/runs": {
      "get": {
        "summary": "Get agent run history",
        "operationId": "getAgentRuns",
        "tags": ["Agents"],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "id",   "in": "path", "required": true, "schema": { "type": "integer" }, "description": "Agent ID" }
        ],
        "responses": {
          "200": {
            "description": "Last 20 runs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "runs": { "type": "array", "items": { "$ref": "#/components/schemas/AgentRun" } } }
                }
              }
            }
          },
          "401": { "description": "Invalid or missing API key", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Agent not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/workspaces/{slug}/agents/{id}/run": {
      "post": {
        "summary": "Run agent (SSE stream)",
        "operationId": "runAgent",
        "tags": ["Agents"],
        "description": "Triggers an agent run and streams output as **Server-Sent Events**.\n\n**Event types:**\n- `{ \"tool_calls\": [\"conn_1_search_issues\"] }` — agent is calling a tool\n- `{ \"done\": true, \"output\": \"...\", \"runId\": 42 }` — run completed\n- `{ \"error\": \"...\" }` — run failed",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "id",   "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": { "type": "string", "description": "Optional instruction to the agent", "example": "Summarise all open Jira tickets assigned to me." }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream of agent output",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string" },
                "example": "data: {\"tool_calls\":[\"conn_1_search_issues\"]}\n\ndata: {\"done\":true,\"output\":\"Found 3 open tickets...\",\"runId\":42}\n\n"
              }
            }
          },
          "401": { "description": "Invalid or missing API key", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Agent not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "429": { "description": "Monthly agent run limit reached", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  }
}